splitting libwallet into internal and external libs
