+++ servers/src/mining/mine_block.rs
@@ -242,7 +242,7 @@ fn get_coinbase(
 			// Get the wallet coinbase
 			let url = format!("{}/v1/receive/coinbase", wallet_listener_url.as_str());
 
-			let res = wallet::client::create_coinbase(&url, &block_fees)?;
+			let res = wallet::libwallet::client::create_coinbase(&url, &block_fees)?;
 			let out_bin = util::from_hex(res.output).unwrap();
 			let kern_bin = util::from_hex(res.kernel).unwrap();
 			let key_id_bin = util::from_hex(res.key_id).unwrap();
