+++ wallet/src/libwallet/internal/restore.rs
