+++ wallet/src/libwallet/internal/sigcontext.rs
