+++ wallet/tests/common/mod.rs
@@ -29,7 +29,7 @@ use core::core::{Output, OutputFeatures, OutputIdentifier, Transaction, TxKernel
 use core::{consensus, global, pow};
 use wallet::file_wallet::*;
 use wallet::libwallet::types::*;
-use wallet::libwallet::updater;
+use wallet::libwallet::internal::updater;
 use wallet::libwallet::{Error, ErrorKind};
 
 use util::secp::pedersen;
@@ -162,7 +162,7 @@ pub fn award_block_to_wallet<T: WalletBackend>(
 		key_id: None,
 		height: prev.height + 1,
 	};
-	let coinbase_tx = wallet::libwallet::updater::receive_coinbase(wallet, &fees);
+	let coinbase_tx = wallet::libwallet::internal::updater::receive_coinbase(wallet, &fees);
 	let (coinbase_tx, fees) = match coinbase_tx {
 		Ok(t) => ((t.0, t.1), t.2),
 		Err(e) => {
